backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.7.2
Pillow==10.4.0
boto3==1.35.32

